Fix 'comparison between signed and unsigned' warnings
The CV builders were incorrectly not checking the correct directories
for warnings in source files under couchstore/, as such a couple of
warnings have snook in. Fix them.

Reviewed-on: http://review.couchbase.org/71912
Tested-by: buildbot <build@couchbase.com>

+++ src/couch_db.cc
@@ -1262,7 +1262,7 @@ couchstore_error_t couchstore_last_os_error(const Db *db,
 
     if (nw < 0) {
         return COUCHSTORE_ERROR_ALLOC_FAIL;
-    } if (nw >= size) {
+    } if (size_t(nw) >= size) {
         /* Truncate the error message */
         buf[size - 1] = '\0';
     }
